#14 callable return type and mock name fixes (#76)
This aligns the callability of return types with the type itself. The type annotations determine whether the returned type is callable or not.

This also fixes a few issues with the mock names that were discovered in the process of writing this.

One of the expect fail tests should be passing by now. #75 was created to represent fixing it.

This addresses #14

There's a notable complexity introduced:

generate_autospec already properly checks callability, but it only works for the item being spec'ed. Child mocks are still callable. This adds logic to enforce the callability of child mocks and also assign a spec to them. The spec is a mock object based on the annotated type. If annotations are missing, then no checks are made.